apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-material-search-api"
version = "0.1.0"
description = "FastAPI backend for Agentic Material Search V2"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "beautifulsoup4>=4.12.3",
  "boto3>=1.35.0",
  "dramatiq[redis]>=1.17.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "lxml>=5.3.0",
  "openai>=1.60.0",
  "pdfplumber>=0.11.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "pypdf>=5.0.1",
  "python-multipart>=0.0.12",
  "sentence-transformers>=3.2.1",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

